# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: PoseFit API
  version: 0.3.0
  description: |
    AI 기반 운동 자세 분석 및 피드백 서비스의 REST API.
    상세 설계 컨텍스트는 본 파일 상단 주석 (1)(2)(3) 참고.
    인증은 소셜 전용(Google). 비밀번호 endpoint 없음.
    응답 필드는 docs/erd.sql (= backend/app/models/) 컬럼과 1:1 정렬 — 예외는 Derived 표시.
  contact:
    name: PoseFit Team
  license:
    name: Proprietary
servers:
  # base path 는 모든 환경에서 /api/v1 로 통일 — 환경 전환 시 host 만 교체.
  #   /api : API 네임스페이스 (FastAPI include_router prefix="/api/v1")
  #   /v1  : API 버전 (breaking change 시 /v2 신설)
  # 아래 도메인은 직관성용 placeholder — 실제 보유/배포 도메인 확정 시 교체.
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://posefit.com/api/v1
    description: Production (placeholder)
security:
  - bearerAuth: []
...
paths:
...
  /exercises:
    get:
      tags: [Exercises]
      operationId: listExercises
      summary: 운동 종목 목록 (Q-R6) — SCR-06
      description: |
        `userAvgScore` 는 호출자 본인의 평균(workout_daily_stats 집계).
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema: { type: boolean, default: true }
          description: false 면 비활성 종목도 포함 (admin 화면 대비).
          example: true
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseListResponse'
              example:
                items:
                  - id: 1
                    nameKo: "스쿼트"
                    nameEn: "Squat"
                    exerciseType: "dynamic"
                    isActive: true
                    userAvgScore: 84.5
                  - id: 2
                    nameKo: "플랭크"
                    nameEn: "Plank"
                    exerciseType: "static"
                    isActive: true
                    userAvgScore: 89.25
                nextCursor: null
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /exercises/{id}/feedbacks:
    get:
      tags: [Feedbacks]
      operationId: listExerciseFeedbacks
      summary: 종목별 당일 세트 피드백 조회 (Q-R10 일부) — 오늘(KST), id ASC
      description: |
        지정한 운동 종목(exerciseId)에 대한 **당일(오늘, KST)** 세트 피드백을 조회한다(이력/표시용).
        운동 중 날짜가 바뀌는 경우는 고려하지 않으며, 조회는 "오늘" 조건만 적용한다. 정렬은 id ASC.

        참고: 이 GET 은 exerciseId 만으로 조회하므로 "오늘 전체"를 돌려준다. 진행 중인
        **이번 묶음(bout)만** 합치는 종합 피드백은 아래 `:summary`(sessionIds 본문)를 사용한다.
      parameters:
        - $ref: '#/components/parameters/ExerciseId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
              example:
                - id: 9001
                  severity: "info"
                  generatedBy: "rule"
                  content: "스쿼트 8회 중 6회는 자세가 매우 안정적이었습니다."
                  createdAt: "2026-06-17T09:47:14Z"
                - id: 9002
                  severity: "warning"
                  generatedBy: "llm"
                  content: "3회차에서 무릎이 발끝보다 앞으로 나왔습니다."
                  createdAt: "2026-06-17T09:47:14Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /exercises/{id}/feedbacks:summary:
    post:
      tags: [Feedbacks]
      operationId: summarizeExerciseFeedbacks
      summary: 운동 종합 피드백 생성 — 이번 묶음 세트 피드백을 LLM 으로 합쳐 1건 반환(미저장)
      description: |
        사용자가 한 운동을 마칠 때 호출한다. 본문으로 이번 운동 묶음(bout)의 **세트 세션 id 목록**
        (sessionIds)을 받아, 그 세션들의 세트 피드백을 모아 RAG/LLM 으로 **종합 피드백 1건**을 만든다.

        - 묶음 범위는 sessionIds 로만 정해진다(현재 묶음만 — "오늘 전체"가 아님).
        - 각 세션은 요청 사용자 소유이고 path 의 exerciseId 와 일치해야 한다. 아니면 검증 실패.
        - 결과는 **저장하지 않고 반환만** 한다. (저장하려면 feedbacks 에 종합/세트 구분 컬럼 등
          ERD 변경이 필요하므로 의도적으로 미저장.)
      parameters:
        - $ref: '#/components/parameters/ExerciseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseFeedbackSummaryRequest'
            example:
              sessionIds: [50220, 50221, 50222]
      responses:
        '200':
          description: OK — 종합 피드백 1건(미저장)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseFeedbackSummaryResponse'
              example:
                exerciseId: 1
                setCount: 3
                generatedBy: "llm"
                content: "총 3세트를 진행했습니다. 전반적으로 깊이는 안정적이었고 세트가 진행될수록 무릎 정렬이 좋아졌습니다. 다만 후반 세트에서 허리가 말리는 경향이 있으니, 다음에는 코어를 단단히 고정한 채 천천히 앉아보세요."
                createdAt: "2026-06-17T10:05:31Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  # ===========================================================================
  # Reports — SCR-05(캘린더), SCR-10
  # ===========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authorization: Bearer <accessToken>.
        access 15분, refresh 14일 (refresh는 별도 endpoint).
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        관리자 Authorization: Bearer <adminAccessToken>.
        소비자 토큰과 JWT type(admin_access/admin_refresh)이 분리되어 상호 사용 불가.
        /admin/* 는 모두 이 스킴으로 보호된다.

  # ===========================================================================
  # Reusable parameters
  # ===========================================================================
  parameters:
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: ["string", "null"]
      description: 직전 응답의 nextCursor. 첫 페이지는 생략.
      example: "eyJpZCI6NTAyMDh9"
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    ExerciseId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: exercises.id (BIGINT).
      example: 1
...
  responses:
    BadRequest:
      description: 잘못된 요청 (스키마/형식 오류)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://posefit.app/errors/bad-request"
            title: "Bad Request"
            status: 400
            code: "BAD_REQUEST"
            detail: "Request body did not match schema."
            errors:
              - field: "exerciseId"
                message: "must be a positive integer"
    Unauthorized:
      description: 인증 필요 또는 토큰 무효
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://posefit.app/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            code: "UNAUTHORIZED"
            detail: "Access token is missing or expired."
...
    NotFound:
      description: 리소스 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://posefit.app/errors/not-found"
            title: "Not Found"
            status: 404
            code: "NOT_FOUND"
            detail: "Workout session 50220 not found."
...
    UnprocessableEntity:
      description: |
        검증 통과했으나 도메인 의미 처리 불가능. 예:
        FACE_NOT_DETECTED / FACE_MISMATCH / FACE_REQUIRED / AGREEMENT_REQUIRED /
        EXERCISE_INACTIVE / DATE_RANGE_IN_FUTURE.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://posefit.app/errors/face-required"
            title: "Face data required"
            status: 422
            code: "FACE_REQUIRED"
            detail: "Register your face before starting a workout."
...
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 Problem Details + 도메인 확장(code, traceId, errors).
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: 오류 종류를 식별하는 URI.
        title:
          type: string
          description: 사람이 읽을 수 있는 짧은 요약.
        status:
          type: integer
          description: HTTP 상태 코드.
        detail:
          type: ["string", "null"]
          description: 사용자에게 보여줄 메시지.
        code:
          $ref: '#/components/schemas/ErrorCode'
        traceId:
          type: ["string", "null"]
          description: 분산 트레이싱 ID (CS 문의 대응용).
        errors:
          type: ["array", "null"]
          description: 필드별 검증 오류 (BadRequest 에 한정).
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    ErrorCode:
      type: string
      description: 도메인 에러 키. 프론트엔드 분기 키로 안정 사용.
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        # 도메인
        - FACE_REQUIRED
        - FACE_ALREADY_REGISTERED
        - FACE_NOT_DETECTED
        - MULTIPLE_FACES_DETECTED
        - FACE_MISMATCH
        - AGREEMENT_REQUIRED
        - SESSION_ALREADY_ACTIVE
        - SESSION_NOT_IN_PROGRESS
        - SESSION_NOT_COMPLETED
        - EXERCISE_INACTIVE
        - LAST_SOCIAL_ACCOUNT
        - SOCIAL_ALREADY_LINKED
        - DATE_RANGE_IN_FUTURE
        - DATE_RANGE_INVALID
        - VIDEO_NOT_UPLOADED
        - VIDEO_TOO_LARGE
        - INVALID_OBJECT_KEY
        - WITHDRAWN_USER

    # ---- Auth -------------------------------------------------------------
...
    ExerciseSummary:
      type: object
      description: 목록용 축약. userAvgScore 는 호출자 본인 평균(derived).
      required: [id, nameKo, exerciseType, isActive]
      properties:
        id:
          type: integer
          format: int64
        nameKo: { type: string }
        nameEn:
          type: ["string", "null"]
        exerciseType:
          type: string
          enum: [static, dynamic]
        isActive:
          type: boolean
        userAvgScore:
          type: ["number", "null"]
          description: Derived — workout_daily_stats 집계 (DECIMAL(5,2)). 미수행 종목은 NULL.
    ExerciseListResponse:
      type: object
      description: 운동 목록 페이지네이션 응답.
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseSummary'
        nextCursor:
          type: ["string", "null"]
...
    Feedback:
      type: object
      description: |
        erd.sql feedbacks 와 1:1. 한 행 = 세트(set) 1회의 피드백. 표시 순서는 id ASC.
        (운동 종합 피드백은 저장하지 않으며 ExerciseFeedbackSummaryResponse 로만 반환된다.)
      required: [id, severity, generatedBy, content, createdAt]
      properties:
        id:
          type: integer
          format: int64
        severity:
          type: string
          enum: [info, warning, critical]
        generatedBy:
          type: string
          enum: [rule, llm]
        content:
          type: string
          description: 사용자에게 보여줄 자연어 피드백.
        createdAt:
          type: string
          format: date-time

    # ---- ExerciseFeedbackSummary (운동 종합 피드백, 미저장) -----------------
    ExerciseFeedbackSummaryRequest:
      type: object
      description: |
        운동을 마칠 때 종합 피드백을 만들 대상 — 이번 묶음(bout)의 세트 세션 id 목록.
        각 세션은 요청 사용자 소유이며 path 의 exerciseId 와 일치해야 한다.
      required: [sessionIds]
      properties:
        sessionIds:
          type: array
          description: 이번 운동에서 진행한 세트 세션 id 들(= 각 STOP 대상 세션). 최소 1개.
          minItems: 1
          items:
            type: integer
            format: int64
    ExerciseFeedbackSummaryResponse:
      type: object
      description: |
        이번 묶음 세트 피드백을 LLM 이 합쳐 만든 운동 종합 피드백 1건. **저장하지 않음**
        (DB 미반영이라 id 가 없다 — 저장하려면 ERD 변경 필요).
      required: [exerciseId, setCount, generatedBy, content, createdAt]
      properties:
        exerciseId:
          type: integer
          format: int64
        setCount:
          type: integer
          description: 종합에 사용된 세트(피드백) 개수.
        generatedBy:
          type: string
          enum: [llm]
          description: 종합 피드백은 항상 LLM 생성.
        content:
          type: string
          description: 여러 세트 피드백을 종합한 운동 전체 코칭 코멘트.
        createdAt:
          type: string
          format: date-time
          description: 생성 시각(서버 응답 시점). 저장하지 않으므로 표시용.

    # ---- Reports ----------------------------------------------------------
...